Change names and structure of hooks.
