+++ src/features/import/components/ImportAlert.tsx
@@ -10,7 +10,7 @@ import {
   Typography,
 } from '@mui/material';
 
-import { Alert as AlertType } from '../hooks/useValidationStep';
+import { Alert as AlertType } from '../hooks/useValidation';
 import messageIds from 'features/import/l10n/messageIds';
 import { useMessages } from 'core/i18n';
